+++ neurons/validator/api/routes/events.py
@@ -1,4 +1,6 @@
-from fastapi import APIRouter, HTTPException
+from typing import Annotated
+
+from fastapi import APIRouter, HTTPException, Query
 
 from neurons.validator.api.types import ApiRequest
 from neurons.validator.db.operations import DatabaseOperations
@@ -6,6 +8,7 @@ from neurons.validator.models.api import (
     GetEventCommunityPrediction,
     GetEventPredictions,
     GetEventResponse,
+    GetEventsCommunityPredictions,
 )
 from neurons.validator.utils.common.interval import get_interval_start_minutes
 
@@ -31,7 +34,7 @@ async def get_event(event_id: str, request: ApiRequest) -> GetEventResponse:
 @router.get(
     "/{event_id}/community_prediction",
 )
-async def get_community_prediction(
+async def get_event_community_prediction(
     event_id: str, request: ApiRequest
 ) -> GetEventCommunityPrediction:
     db_operations: DatabaseOperations = request.state.db_operations
@@ -45,11 +48,47 @@ async def get_community_prediction(
 
     current_interval = get_interval_start_minutes()
 
-    community_prediction = await db_operations.get_wa_prediction_event(
-        unique_event_id=unique_event_id, interval_start_minutes=current_interval
+    community_predictions_by_id = await db_operations.get_wa_predictions_events(
+        unique_event_ids=[unique_event_id], interval_start_minutes=current_interval
     )
 
-    return {"event_id": event_id, "community_prediction": community_prediction}
+    return {
+        "event_id": unique_event_id.removeprefix("ifgames-"),
+        "community_prediction": community_predictions_by_id.get(unique_event_id),
+    }
+
+
+@router.get(
+    "/community_predictions/",
+)
+async def get_events_community_predictions(
+    event_ids: Annotated[list[str], Query()], request: ApiRequest
+) -> GetEventsCommunityPredictions:
+    db_operations: DatabaseOperations = request.state.db_operations
+
+    if len(event_ids) > 20:
+        raise HTTPException(
+            status_code=400,
+            detail="Maximum 20 events ids allowed",
+        )
+
+    current_interval = get_interval_start_minutes()
+
+    unique_event_ids = [f"ifgames-{event_id}" for event_id in event_ids]
+
+    community_predictions_by_id = await db_operations.get_wa_predictions_events(
+        unique_event_ids=unique_event_ids, interval_start_minutes=current_interval
+    )
+
+    community_predictions = [
+        {
+            "event_id": unique_event_id.removeprefix("ifgames-"),
+            "community_prediction": community_predictions_by_id[unique_event_id],
+        }
+        for unique_event_id in community_predictions_by_id.keys()
+    ]
+
+    return {"count": len(community_predictions), "community_predictions": community_predictions}
 
 
 @router.get(