+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Release Notes
 
+## [1.4.6] - 2025-04-29
+- **Scoring**: Improved metagraph scoring to correct for class imbalance by up-weighting less frequent YES outcomes
+- **Database**: Added task to delete processed old scores
+- **API**: Batched community predictions API
+
 ## [1.4.5] - 2025-04-22
 - **Bittensor Upgrade**: Upgraded to Bittensor version 9.4.0